Renamed `@sentry.report_exceptions` => `@sentry.report_exceptions`
And also moved common modules' code in `__init__.py` files

As requested in PR: https://github.com/ministryofjustice/analytics-platform-ops/pull/83
